fix: uses parseFloat to handle borderWidth between 0 and 1
Sometimes depending on the browser zoom, the borderWidth can be between 0 and 1. Because of the parseInt, in these cases, the value was hidden, even though it was greater than 0.

Example, with a zoom of 150% and a borderWidth of 1px, the rendered borderWidth is 0.667px.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1732769

+++ app/utilities/styles.js
@@ -6,7 +6,7 @@ export const getStyle = (el, name) => {
     name = name.toLowerCase()
     let s = document.defaultView.getComputedStyle(el, '')
     return s && s.getPropertyValue(name)
-  } 
+  }
 }
 
 export const getStyles = el => {
@@ -30,7 +30,7 @@ export const getStyles = el => {
 
   const { borderColor, borderWidth, borderStyle } = borders
 
-  if (parseInt(borderWidth) > 0) {
+  if (parseFloat(borderWidth) > 0) {
     vettedStyles.push({
       prop: 'borderColor',
       value: borderColor,
@@ -107,7 +107,7 @@ export const findNearestParentElement = el => {
 export const findNearestChildElement = el => {
   if (el.shadowRoot && el.shadowRoot.children.length) {
     return [...el.shadowRoot.children]
-      .filter(({nodeName}) => 
+      .filter(({nodeName}) =>
         !['LINK','STYLE','SCRIPT','HTML','HEAD'].includes(nodeName)
       )[0]
   }
@@ -120,7 +120,7 @@ export const loadStyles = async stylesheets => {
   const texts   = await Promise.all(fetches.map(url => url.text()))
   const style   = document.createElement('style')
 
-  style.textContent = texts.reduce((styles, fileContents) => 
+  style.textContent = texts.reduce((styles, fileContents) =>
     styles + fileContents
   , '')
 
@@ -156,8 +156,8 @@ export const firstUsableFontFromFamily = family => {
 
       context.font = `12px ${fontName}, sans-serif`
 
-      return baselineSize !== context.measureText('font-test').width 
-        ? fontName 
+      return baselineSize !== context.measureText('font-test').width
+        ? fontName
         : false
   }).filter(value => value !== false)[0]
